code reformated into functions, custom plot is depricated
