+++ visualisation/__init__.py
@@ -1,2 +1,3 @@
 from . import custom_plot
 from . import draw
+from . import visualisation_options
